# docs/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Fisco.io - REST API
  version: 0.1.0
  description: |
    API REST del framework de administración tributaria Fisco.io.
    Comandos (POST) y consultas (GET) sobre sujetos y obligaciones.
    Actualizar este spec cuando se agregue o modifique un endpoint.
servers:
  - url: /api
    description: API base path (relative to host)
paths:
...
  /obligations/{obligation_id}/liquidations:
    post:
      summary: Crear liquidación
      description: Comando CreateLiquidation. Registra una liquidación para la obligación en un período.
      operationId: createLiquidation
      tags: [Obligations]
      parameters:
        - name: obligation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLiquidationRequest"
      responses:
        "201":
          description: Liquidación creada
        "400":
          description: Parámetros inválidos
        "404":
          description: Obligación no encontrada
...
components:
  schemas:
...
    CreateLiquidationRequest:
      type: object
      required: [period, tax_amount]
      properties:
        period:
          type: string
          description: Período fiscal (ej. 2024-01)
        tax_amount:
          type: number
          format: double
...